vfio/common: Change vfio_devices_all_running_and_saving() logic to equivalent one
vfio_devices_all_running_and_saving() is used to check if migration is
in pre-copy phase. This is done by checking if migration is in setup or
active states and if all VFIO devices are in pre-copy state, i.e.
_SAVING | _RUNNING.

VFIO migration protocol v2 currently doesn't support pre-copy phase, so
it doesn't have an equivalent VFIO pre-copy state like v1 has.

As preparation for adding the v2 protocol, change
vfio_devices_all_running_and_saving() logic to check if migration
is active [*] and if all VFIO devices are in running state.

The new logic avoids using the VFIO pre-copy state and is equivalent to
the previous one, indicating that migration is in pre-copy phase. No
functional changes intended.

[*] Note that checking if migration is setup or active and if all VFIO
devices are running doesn't guarantee that we are in pre-copy phase,
thus we check if migration is only active.

Signed-off-by: Avihai Horon <avihaih@nvidia.com>