vfio/common: Record DMA mapped ranges
According to the device DMA logging uAPI, DMA ranges to be logged by
the device must be provided all at once upon DMA logging start.

This means that when a range is DMA mapped, it should be recorded so
later it can be used for DMA logging start. Currently, this is not the
case and the DMA mapped ranges are not recorded.

Hence, as preparation for adding VFIO devices DMA logging, record the
ranges that were DMA mapped.

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Signed-off-by: Avihai Horon <avihaih@nvidia.com>