vfio: Make VFIO migration non-experimental
The major parts of VFIO migration are supported today in QEMU. This
includes basic VFIO migration, device dirty page tracking and VFIO
migration precopy support.

Thus, at this point in time, it seems appropriate to make VFIO migration
non-experimental. Remove the x prefix from enable_migration property,
change it to ON_OFF_AUTO and let the default value be AUTO.

When enable_migration is not explicitly enabled (i.e., its value is
AUTO), migration will be blocked if device dirty page tracking is not
supported by the device. This is because currently this is the only
method a VFIO device can do dirty page tracking, which is essential for
migrating in a reasonable downtime. Setting enable_migration to ON will
allow migration even when device dirty page tracking is not supported by
the device, and in this case the entire DMA-able RAM by the device will
be perpetually marked dirty.

While at it, remove a redundant assignment in vfio_migration_realize().

Signed-off-by: Avihai Horon <avihaih@nvidia.com>
Change-Id: I8f481aa9e6400d54c9d8608c57ab950655240911