migration: Implement pre-copy init logic
This patch implements the logic for the new pre-copy init capability.

A new migration command is added, MIG_CMD_PRECOPY_INIT_ENABLE, to notify
the destination that pre-copy init is used, as well as a new return path
message, MIG_RP_MSG_PRECOPY_INIT_LOADED, to notify the source that
pre-copy initial data has been loaded.

In addition, two new handlers are added to SaveVMHandlers,
precopy_init_supported and precopy_init_loaded, which are called in the
destination and indicate if pre-copy init is supported and if pre-copy
initial data has been loaded, respectively.

Using the above, when pre-copy init capability is enabled, the migration
will allow sending of pre-copy initial data and will not stop the source
VM and complete migration until an ACK has been received from the
destination that the pre-copy initial data has been loaded.

Note that it's the migration clients' responsibility to notify their
counterparts in the destination that their initial pre-copy data has
been sent.

Signed-off-by: Avihai Horon <avihaih@nvidia.com>
Change-Id: I83942cde3d26610ada8c96d80a29c700dd665add