migration: Add pre-copy init capability
Currently, migration downtime and threshold estimations are calculated
based on bandwidth and remaining migration data to be transferred. This
assumes that loading the migration data in the destination takes only a
negligible amount of time.

While this may be true for RAM, it's not necessarily true for other
migration clients. For example, loading the data of a VFIO device in the
destination might also require from the device to allocate resources,
prepare internal data structures and so on. These operations can take a
significant amount of time which can increase migration downtime.

This patch adds a new capability "pre-copy init" that allows the source
side to send initial pre-copy data and the destination side to ACK that
this data has been loaded. The source side will not attempt to stop the
VM and complete the migration until this ACK is received.

This way migration clients can send their initial pre-copy data which
can be used by the destination to pre-allocate all needed resources, and
thus prevent these initializations from happening during downtime.

This new capability relies on the return path capability to communicate
from the destination to the source.

The actual implementation of the capability logic will be added in the
following patches.

Signed-off-by: Avihai Horon <avihaih@nvidia.com>
Change-Id: I8170e45f9a1b78c0e118df93de7a0f0a2c00195a