vfio/migration: Move from STOP_COPY to STOP in vfio_save_cleanup()
Changing device state from STOP_COPY to STOP can take time as the device
might need to free resources etc. Currently, this is done in
vfio_save_complete_precopy() and therefore is counted in the migration
downtime.

To avoid this and reduce migration downtime, change the device state
from STOP_COPY to STOP in vfio_save_cleanup() which is not part of
migration downtime.

Signed-off-by: Avihai Horon <avihaih@nvidia.com>
Change-Id: I37b6a30dfed11c79e65a837fcf4b42e6432eaa84